namespaces: Create a new per namespace process library
The goal of this library is to be able to inject per namespace
variables and code in the process address space.
This way a new namespace process will have his own customized
version of the .so variable and the right code for processing
them.

This commit starts by applyings this .so trick to environment
variables.

The idea comes from an idea I had that Nadav drastically improved
on the WP4 mailling list.

Signed-off-by: Benoît Canet <benoit@scylladb.com>