nfs: Add nfs server
This NFS server will be used for the unit tests.

It's done this way because nfs-kernel-server is broken
on some distribution due to systemd introduction without
corresponding unit files.

nfs-ganesha was not usable because it is not currently
packaged for debian.

The nfs-server target will build external/fs/unfsd.bin
when the user will need to run the test.

Signed-off-by: Benoît Canet <benoit@scylladb.com>