mmu: Make file_vma::sync flush dirty pages to disk
NFS not having a cache vfs_file::mmap() will construct and
return a mmu::default_file_mmap object that does not implement
writes.

The mmu code will try to flush the dirty pages when msync() or
munmap() is called but until now it does not write the dirty
pages to disk before calling the flush vfs method.

Implement this write only for the dirty pages.

Signed-off-by: Benoît Canet <benoit@scylladb.com>