vfs: Release/Get the vnode lock around read/write VOPs
The very coarse grained locking around read/write VOPs
have for consequence to serialize all VFS operations
to the same file in these calls.

This will for sure kill every bit of multithreaded
performance when accessing a single file.

This patch release and retake the lock around these
VFS operations and take care of taking the vnode lock
inside the filesystems that are not multithread ready.
(every fs excepted ZFS).

ZFS in turn use a range lock around it's internal structure
so it should be OK.

Signed-off-by: Benoît Canet <benoit@scylladb.com>