app: Factorize argc and argv construction
The go c-shared object need the libc to pass them
the argc and argv parameters in order to initialize
the runtime properly.

This is a preparation patch for this.

Since I like to be able to seemlingly prove code move
in my head this make use of a lot of class variables
that I will get rid of in next patches.

Signed-off-by: Benoît Canet <benoit@scylladb.com>