vdso: vdso fast syscall compatibility library
Go use this on linux to lookup time. We need a shared
libary that we will expose to the go runtime by passing
it's pointer via auxv in order for go to resolve this
kinds of symbols.

We cannot pass the kernel library as vdso because it's
located at address 0 and Go runtime will guard against
this.

Signed-off-by: Benoît Canet <benoit@scylladb.com>