app: Use GoMain as go entry point
Go c-shared library main() get ignored by the
Go runtime at init. main() cannot be //export since the
c-go .c file already have a main. Fact is that if main()
is exported in a Go "C" module the C compiler will complain
that two main are presents.

So this patch declare GoMain as the official OSv go main()
entry point. Since the go library init already took argc and
argv as parameters GoMain does not need to bother with them
and hence is specific to Go.:

Signed-off-by: Benoît Canet <benoit@scylladb.com>